src/main/java/com/ehr/patient/service/impl/PatientServiceImpl.java
package com.ehr.patient.service.impl;

import com.ehr.patient.domain.Patient;
import com.ehr.patient.repository.PatientRepository;
import com.ehr.patient.service.PatientService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;
import java.util.Optional;

@Service
@Transactional
public class PatientServiceImpl implements PatientService {

    @Autowired
    private PatientRepository patientRepository;

    @Override
    public Patient save(Patient patient) {
        return patientRepository.save(patient);
    }

    @Override
    public List<Patient> findAll() {
        return patientRepository.findAll();
    }

    @Override
    public Optional<Patient> getOne(Long id) {
        return patientRepository.findById(id);
    }
}
